multi connection add 및 암호화 config common 이동
