+++ bkjeon-mybatis-codebase/base-common/src/main/java/com/example/bkjeon/config/mybatis/OracleDataSourceConfig.java
@@ -1,4 +1,4 @@
-package com.example.bkjeon.mybatis;
+package com.example.bkjeon.config.mybatis;
 
 import javax.sql.DataSource;
 import org.apache.ibatis.session.SqlSessionFactory;
